src/DetectSegment/requirements.txt
torch
transformers
accelerate
safetensors
Pillow
numpy
opencv-python
segment-anything
